add data in seeddata file

+++ Models/Spoons.cs
@@ -7,12 +7,13 @@ namespace MvcSpoon.Models
 {
     public class Spoons
     {
+        // Add a field with datatype
         public int Id { get; set; }
         public string Type { get; set; }
         public string Size { get; set; }
         public decimal Price { get; set; }
         public string Material { get; set; }
         public string Use { get; set; }
-        public int Rating { get; set; }
+        public float Rating { get; set; }
     }
 }
